@@ -1,150 +1,292 @@
 const card = {
-    "object":"card",
-    "id":"ac709474-7790-483f-9ed5-ea7abccfce53",
-    "oracle_id":"b2c6aa39-2d2a-459c-a555-fb48ba993373",
-    "multiverse_ids":[],
-    "name":"Island",
-    "lang":"en",
-    "released_at":"2019-02-16",
-    "uri":"https://api.scryfall.com/cards/ac709474-7790-483f-9ed5-ea7abccfce53",
-    "scryfall_uri":"https://scryfall.com/card/prw2/10/island?utm_source=api",
-    "layout":"normal",
-    "highres_image":false,
-    "image_uris":{
-        "small":"https://img.scryfall.com/cards/small/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699",
-        "normal":"https://img.scryfall.com/cards/normal/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699",
-        "large":"https://img.scryfall.com/cards/large/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699",
-        "png":"https://img.scryfall.com/cards/png/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.png?1545156699",
-        "art_crop":"https://img.scryfall.com/cards/art_crop/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699",
-        "border_crop":"https://img.scryfall.com/cards/border_crop/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699"
-    },
-    "mana_cost":"",
-    "cmc":0.0,
-    "type_line":"Basic Land — Island",
-    "oracle_text":"({T}: Add {U}.)",
-    "colors":[],
-    "color_identity":["U"],
-    "legalities":{
-        "standard":"legal",
-        "future":"legal",
-        "frontier":"legal",
-        "modern":"legal",
-        "legacy":"legal",
-        "pauper":"legal",
-        "vintage":"legal",
-        "penny":"legal",
-        "commander":"legal",
-        "1v1":"legal",
-        "duel":"legal",
-        "brawl":"legal"
-    },
-    "games":["paper"],
-    "reserved":false,
-    "foil":true,
-    "nonfoil":false,
-    "oversized":false,
-    "promo":true,
-    "reprint":true,
-    "set":"prw2",
-    "set_name":"RNA Ravnica Weekend",
-    "set_uri":"https://api.scryfall.com/sets/ee3a8eb6-0583-492b-8be5-265795d38038",
-    "set_search_uri":"https://api.scryfall.com/cards/search?order=set\u0026q=e%3Aprw2\u0026unique=prints",
-    "scryfall_set_uri":"https://scryfall.com/sets/prw2?utm_source=api",
-    "rulings_uri":"https://api.scryfall.com/cards/ac709474-7790-483f-9ed5-ea7abccfce53/rulings",
-    "prints_search_uri":"https://api.scryfall.com/cards/search?order=released\u0026q=oracleid%3Ab2c6aa39-2d2a-459c-a555-fb48ba993373\u0026unique=prints",
-    "collector_number":"10",
-    "digital":false,
-    "rarity":"common",
-    "watermark":"simic",
-    "illustration_id":"157dfb2c-428b-41a6-aead-f8809291957d",
-    "artist":"Yeong-Hao Han",
-    "border_color":"black",
-    "frame":"2015",
-    "frame_effect":"",
-    "full_art":false,
-    "story_spotlight":false,
-    "related_uris":{
-        "tcgplayer_decks":"https://decks.tcgplayer.com/magic/deck/search?contains=Island\u0026page=1\u0026partner=Scryfall\u0026utm_campaign=affiliate\u0026utm_medium=scryfall\u0026utm_source=scryfall",
-        "edhrec":"http://edhrec.com/route/?cc=Island",
-        "mtgtop8":"http://mtgtop8.com/search?MD_check=1\u0026SB_check=1\u0026cards=Island"
-    }
+  object: 'card',
+  id: 'ac709474-7790-483f-9ed5-ea7abccfce53',
+  oracle_id: 'b2c6aa39-2d2a-459c-a555-fb48ba993373',
+  multiverse_ids: [],
+  name: 'Island',
+  lang: 'en',
+  released_at: '2019-02-16',
+  uri: 'https://api.scryfall.com/cards/ac709474-7790-483f-9ed5-ea7abccfce53',
+  scryfall_uri: 'https://scryfall.com/card/prw2/10/island?utm_source=api',
+  layout: 'normal',
+  highres_image: false,
+  image_uris: {
+    small:
+      'https://img.scryfall.com/cards/small/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699',
+    normal:
+      'https://img.scryfall.com/cards/normal/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699',
+    large:
+      'https://img.scryfall.com/cards/large/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699',
+    png: 'https://img.scryfall.com/cards/png/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.png?1545156699',
+    art_crop:
+      'https://img.scryfall.com/cards/art_crop/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699',
+    border_crop:
+      'https://img.scryfall.com/cards/border_crop/front/a/c/ac709474-7790-483f-9ed5-ea7abccfce53.jpg?1545156699',
+  },
+  mana_cost: '',
+  cmc: 0.0,
+  type_line: 'Basic Land — Island',
+  oracle_text: '({T}: Add {U}.)',
+  colors: [],
+  color_identity: ['U'],
+  legalities: {
+    standard: 'legal',
+    future: 'legal',
+    frontier: 'legal',
+    modern: 'legal',
+    legacy: 'legal',
+    pauper: 'legal',
+    vintage: 'legal',
+    penny: 'legal',
+    commander: 'legal',
+    '1v1': 'legal',
+    duel: 'legal',
+    brawl: 'legal',
+  },
+  games: ['paper'],
+  reserved: false,
+  foil: true,
+  nonfoil: false,
+  oversized: false,
+  promo: true,
+  reprint: true,
+  set: 'prw2',
+  set_name: 'RNA Ravnica Weekend',
+  set_uri: 'https://api.scryfall.com/sets/ee3a8eb6-0583-492b-8be5-265795d38038',
+  set_search_uri:
+    'https://api.scryfall.com/cards/search?order=set\u0026q=e%3Aprw2\u0026unique=prints',
+  scryfall_set_uri: 'https://scryfall.com/sets/prw2?utm_source=api',
+  rulings_uri:
+    'https://api.scryfall.com/cards/ac709474-7790-483f-9ed5-ea7abccfce53/rulings',
+  prints_search_uri:
+    'https://api.scryfall.com/cards/search?order=released\u0026q=oracleid%3Ab2c6aa39-2d2a-459c-a555-fb48ba993373\u0026unique=prints',
+  collector_number: '10',
+  digital: false,
+  rarity: 'common',
+  watermark: 'simic',
+  illustration_id: '157dfb2c-428b-41a6-aead-f8809291957d',
+  artist: 'Yeong-Hao Han',
+  border_color: 'black',
+  frame: '2015',
+  frame_effect: '',
+  full_art: false,
+  story_spotlight: false,
+  related_uris: {
+    tcgplayer_decks:
+      'https://decks.tcgplayer.com/magic/deck/search?contains=Island\u0026page=1\u0026partner=Scryfall\u0026utm_campaign=affiliate\u0026utm_medium=scryfall\u0026utm_source=scryfall',
+    edhrec: 'http://edhrec.com/route/?cc=Island',
+    mtgtop8:
+      'http://mtgtop8.com/search?MD_check=1\u0026SB_check=1\u0026cards=Island',
+  },
 }
 
-
-
 const card2 = {
-    "object":"card",
-    "id":"09635727-dd96-4a4f-b225-e410c4053499",
-    "oracle_id":"8314bdf5-1539-4215-949e-f46584cb43bb",
-    "multiverse_ids":[159093],
-    "mtgo_id":28113,
-    "mtgo_foil_id":28114,
-    "name":"Gargantuan Gorilla",
-    "lang":"en",
-    "released_at":"2007-09-10",
-    "uri":"https://api.scryfall.com/cards/09635727-dd96-4a4f-b225-e410c4053499",
-    "scryfall_uri":"https://scryfall.com/card/me1/119/gargantuan-gorilla?utm_source=api",
-    "layout":"normal",
-    "highres_image":true,
-    "image_uris":{
-        "small":"https://img.scryfall.com/cards/small/en/me1/119.jpg?1535856113",
-        "normal":"https://img.scryfall.com/cards/normal/en/me1/119.jpg?1535856113",
-        "large":"https://img.scryfall.com/cards/large/en/me1/119.jpg?1535856113",
-        "png":"https://img.scryfall.com/cards/png/en/me1/119.png?1535856113",
-        "art_crop":"https://img.scryfall.com/cards/art_crop/en/me1/119.jpg?1535856113",
-        "border_crop":"https://img.scryfall.com/cards/border_crop/en/me1/119.jpg?1535856113"
-    },
-    "mana_cost":"{4}{G}{G}{G}",
-    "cmc":7.0,
-    "type_line":"Creature — Ape",
-    "oracle_text":"At the beginning of your upkeep, you may sacrifice a Forest. If you sacrifice a snow Forest this way, Gargantuan Gorilla gains trample until end of turn. If you don't sacrifice a Forest, sacrifice Gargantuan Gorilla and it deals 7 damage to you.\n{T}: Gargantuan Gorilla deals damage equal to its power to another target creature. That creature deals damage equal to its power to Gargantuan Gorilla.",
-    "power":"7",
-    "toughness":"7",
-    "colors":["G"],
-    "color_identity":["G"],
-    "legalities":{"standard":"not_legal",
-    "future":"not_legal",
-    "frontier":"not_legal",
-    "modern":"not_legal",
-    "legacy":"legal",
-    "pauper":"not_legal",
-    "vintage":"legal",
-    "penny":"legal",
-    "commander":"legal",
-    "1v1":"legal",
-    "duel":"legal",
-    "brawl":"not_legal"},
-    "games":["mtgo"],
-    "reserved":true,
-    "foil":true,
-    "nonfoil":true,
-    "oversized":false,
-    "promo":false,
-    "reprint":true,
-    "set":"me1",
-    "set_name":"Masters Edition",
-    "set_uri":"https://api.scryfall.com/sets/407d388d-1abf-4c1d-b0c6-f56280898a1a",
-    "set_search_uri":"https://api.scryfall.com/cards/search?order=set\u0026q=e%3Ame1\u0026unique=prints",
-    "scryfall_set_uri":"https://scryfall.com/sets/me1?utm_source=api",
-    "rulings_uri":"https://api.scryfall.com/cards/09635727-dd96-4a4f-b225-e410c4053499/rulings",
-    "prints_search_uri":"https://api.scryfall.com/cards/search?order=released\u0026q=oracleid%3A8314bdf5-1539-4215-949e-f46584cb43bb\u0026unique=prints",
-    "collector_number":"119",
-    "digital":true,
-    "rarity":"rare",
-    "illustration_id":"bd2152f6-b208-4104-9802-a4621f4bc50e",
-    "artist":"Greg Simanson",
-    "border_color":"black",
-    "frame":"1997",
-    "frame_effect":"",
-    "full_art":false,
-    "story_spotlight":false,
-    "edhrec_rank":16799,
-    "related_uris":{
-        "gatherer":"http://gatherer.wizards.com/Pages/Card/Details.aspx?multiverseid=159093",
-        "tcgplayer_decks":"https://decks.tcgplayer.com/magic/deck/search?contains=Gargantuan+Gorilla\u0026page=1\u0026partner=Scryfall\u0026utm_campaign=affiliate\u0026utm_medium=scryfall\u0026utm_source=scryfall",
-        "edhrec":"http://edhrec.com/route/?cc=Gargantuan+Gorilla",
-        "mtgtop8":"http://mtgtop8.com/search?MD_check=1\u0026SB_check=1\u0026cards=Gargantuan+Gorilla"
-    }
+  object: 'card',
+  id: '09635727-dd96-4a4f-b225-e410c4053499',
+  oracle_id: '8314bdf5-1539-4215-949e-f46584cb43bb',
+  multiverse_ids: [159093],
+  mtgo_id: 28113,
+  mtgo_foil_id: 28114,
+  name: 'Gargantuan Gorilla',
+  lang: 'en',
+  released_at: '2007-09-10',
+  uri: 'https://api.scryfall.com/cards/09635727-dd96-4a4f-b225-e410c4053499',
+  scryfall_uri:
+    'https://scryfall.com/card/me1/119/gargantuan-gorilla?utm_source=api',
+  layout: 'normal',
+  highres_image: true,
+  image_uris: {
+    small: 'https://img.scryfall.com/cards/small/en/me1/119.jpg?1535856113',
+    normal: 'https://img.scryfall.com/cards/normal/en/me1/119.jpg?1535856113',
+    large: 'https://img.scryfall.com/cards/large/en/me1/119.jpg?1535856113',
+    png: 'https://img.scryfall.com/cards/png/en/me1/119.png?1535856113',
+    art_crop:
+      'https://img.scryfall.com/cards/art_crop/en/me1/119.jpg?1535856113',
+    border_crop:
+      'https://img.scryfall.com/cards/border_crop/en/me1/119.jpg?1535856113',
+  },
+  mana_cost: '{4}{G}{G}{G}',
+  cmc: 7.0,
+  type_line: 'Creature — Ape',
+  oracle_text:
+    "At the beginning of your upkeep, you may sacrifice a Forest. If you sacrifice a snow Forest this way, Gargantuan Gorilla gains trample until end of turn. If you don't sacrifice a Forest, sacrifice Gargantuan Gorilla and it deals 7 damage to you.\n{T}: Gargantuan Gorilla deals damage equal to its power to another target creature. That creature deals damage equal to its power to Gargantuan Gorilla.",
+  power: '7',
+  toughness: '7',
+  colors: ['G'],
+  color_identity: ['G'],
+  legalities: {
+    standard: 'not_legal',
+    future: 'not_legal',
+    frontier: 'not_legal',
+    modern: 'not_legal',
+    legacy: 'legal',
+    pauper: 'not_legal',
+    vintage: 'legal',
+    penny: 'legal',
+    commander: 'legal',
+    '1v1': 'legal',
+    duel: 'legal',
+    brawl: 'not_legal',
+  },
+  games: ['mtgo'],
+  reserved: true,
+  foil: true,
+  nonfoil: true,
+  oversized: false,
+  promo: false,
+  reprint: true,
+  set: 'me1',
+  set_name: 'Masters Edition',
+  set_uri: 'https://api.scryfall.com/sets/407d388d-1abf-4c1d-b0c6-f56280898a1a',
+  set_search_uri:
+    'https://api.scryfall.com/cards/search?order=set\u0026q=e%3Ame1\u0026unique=prints',
+  scryfall_set_uri: 'https://scryfall.com/sets/me1?utm_source=api',
+  rulings_uri:
+    'https://api.scryfall.com/cards/09635727-dd96-4a4f-b225-e410c4053499/rulings',
+  prints_search_uri:
+    'https://api.scryfall.com/cards/search?order=released\u0026q=oracleid%3A8314bdf5-1539-4215-949e-f46584cb43bb\u0026unique=prints',
+  collector_number: '119',
+  digital: true,
+  rarity: 'rare',
+  illustration_id: 'bd2152f6-b208-4104-9802-a4621f4bc50e',
+  artist: 'Greg Simanson',
+  border_color: 'black',
+  frame: '1997',
+  frame_effect: '',
+  full_art: false,
+  story_spotlight: false,
+  edhrec_rank: 16799,
+  related_uris: {
+    gatherer:
+      'http://gatherer.wizards.com/Pages/Card/Details.aspx?multiverseid=159093',
+    tcgplayer_decks:
+      'https://decks.tcgplayer.com/magic/deck/search?contains=Gargantuan+Gorilla\u0026page=1\u0026partner=Scryfall\u0026utm_campaign=affiliate\u0026utm_medium=scryfall\u0026utm_source=scryfall',
+    edhrec: 'http://edhrec.com/route/?cc=Gargantuan+Gorilla',
+    mtgtop8:
+      'http://mtgtop8.com/search?MD_check=1\u0026SB_check=1\u0026cards=Gargantuan+Gorilla',
+  },
+}
+
+const card3 = {
+  object: 'card',
+  id: '000366c8-7a43-49d7-a103-ac5bd7efd9aa',
+  oracle_id: '56719f6a-1a6c-4c0a-8d21-18f7d7350b68',
+  multiverse_ids: [4925],
+  mtgo_id: 9701,
+  mtgo_foil_id: 9702,
+  tcgplayer_id: 18500,
+  cardmarket_id: 9070,
+  name: 'Swamp',
+  lang: 'en',
+  released_at: '1997-10-14',
+  uri: 'https://api.scryfall.com/cards/000366c8-7a43-49d7-a103-ac5bd7efd9aa',
+  scryfall_uri: 'https://scryfall.com/card/tmp/340/swamp?utm_source=api',
+  layout: 'normal',
+  highres_image: true,
+  image_status: 'highres_scan',
+  image_uris: {
+    small:
+      'https://cards.scryfall.io/small/front/0/0/000366c8-7a43-49d7-a103-ac5bd7efd9aa.jpg?1562052318',
+    normal:
+      'https://cards.scryfall.io/normal/front/0/0/000366c8-7a43-49d7-a103-ac5bd7efd9aa.jpg?1562052318',
+    large:
+      'https://cards.scryfall.io/large/front/0/0/000366c8-7a43-49d7-a103-ac5bd7efd9aa.jpg?1562052318',
+    png: 'https://cards.scryfall.io/png/front/0/0/000366c8-7a43-49d7-a103-ac5bd7efd9aa.png?1562052318',
+    art_crop:
+      'https://cards.scryfall.io/art_crop/front/0/0/000366c8-7a43-49d7-a103-ac5bd7efd9aa.jpg?1562052318',
+    border_crop:
+      'https://cards.scryfall.io/border_crop/front/0/0/000366c8-7a43-49d7-a103-ac5bd7efd9aa.jpg?1562052318',
+  },
+  mana_cost: '',
+  cmc: 0.0,
+  type_line: 'Basic Land — Swamp',
+  oracle_text: '({T}: Add {B}.)',
+  colors: [],
+  color_identity: ['B'],
+  keywords: [],
+  produced_mana: ['B'],
+  legalities: {
+    standard: 'legal',
+    future: 'legal',
+    historic: 'legal',
+    gladiator: 'legal',
+    pioneer: 'legal',
+    explorer: 'legal',
+    modern: 'legal',
+    legacy: 'legal',
+    pauper: 'legal',
+    vintage: 'legal',
+    penny: 'legal',
+    commander: 'legal',
+    oathbreaker: 'legal',
+    brawl: 'legal',
+    historicbrawl: 'legal',
+    alchemy: 'legal',
+    paupercommander: 'legal',
+    duel: 'legal',
+    oldschool: 'not_legal',
+    premodern: 'legal',
+    predh: 'legal',
+  },
+  games: ['paper', 'mtgo'],
+  reserved: false,
+  foil: false,
+  nonfoil: true,
+  finishes: ['nonfoil'],
+  oversized: false,
+  promo: false,
+  reprint: true,
+  variation: false,
+  set_id: '10df3a67-178e-4363-8668-34f0e6edf2a7',
+  set: 'tmp',
+  set_name: 'Tempest',
+  set_type: 'expansion',
+  set_uri: 'https://api.scryfall.com/sets/10df3a67-178e-4363-8668-34f0e6edf2a7',
+  set_search_uri:
+    'https://api.scryfall.com/cards/search?order=set\u0026q=e%3Atmp\u0026unique=prints',
+  scryfall_set_uri: 'https://scryfall.com/sets/tmp?utm_source=api',
+  rulings_uri:
+    'https://api.scryfall.com/cards/000366c8-7a43-49d7-a103-ac5bd7efd9aa/rulings',
+  prints_search_uri:
+    'https://api.scryfall.com/cards/search?order=released\u0026q=oracleid%3A56719f6a-1a6c-4c0a-8d21-18f7d7350b68\u0026unique=prints',
+  collector_number: '340',
+  digital: false,
+  rarity: 'common',
+  card_back_id: '0aeebaf5-8c7d-4636-9e82-8c27447861f7',
+  artist: 'Brom',
+  artist_ids: ['bd79c951-2552-45dd-aae7-a333c2b7c1ef'],
+  illustration_id: 'bf15950f-cff4-442c-ab48-96f2042d49a2',
+  border_color: 'black',
+  frame: '1997',
+  full_art: false,
+  textless: false,
+  booster: true,
+  story_spotlight: false,
+  prices: {
+    usd: '0.37',
+    usd_foil: null,
+    usd_etched: null,
+    eur: '0.50',
+    eur_foil: null,
+    tix: '0.04',
+  },
+  related_uris: {
+    gatherer:
+      'https://gatherer.wizards.com/Pages/Card/Details.aspx?multiverseid=4925',
+    tcgplayer_infinite_articles:
+      'https://infinite.tcgplayer.com/search?contentMode=article\u0026game=magic\u0026partner=scryfall\u0026q=Swamp\u0026utm_campaign=affiliate\u0026utm_medium=api\u0026utm_source=scryfall',
+    tcgplayer_infinite_decks:
+      'https://infinite.tcgplayer.com/search?contentMode=deck\u0026game=magic\u0026partner=scryfall\u0026q=Swamp\u0026utm_campaign=affiliate\u0026utm_medium=api\u0026utm_source=scryfall',
+    edhrec: 'https://edhrec.com/route/?cc=Swamp',
+  },
+  purchase_uris: {
+    tcgplayer:
+      'https://www.tcgplayer.com/product/18500?page=1\u0026utm_campaign=affiliate\u0026utm_medium=api\u0026utm_source=scryfall',
+    cardmarket:
+      'https://www.cardmarket.com/en/Magic/Products/Search?referrer=scryfall\u0026searchString=Swamp\u0026utm_campaign=card_prices\u0026utm_medium=text\u0026utm_source=scryfall',
+    cardhoarder:
+      'https://www.cardhoarder.com/cards/9701?affiliate_id=scryfall\u0026ref=card-profile\u0026utm_campaign=affiliate\u0026utm_medium=card\u0026utm_source=scryfall',
+  },
 }
 
 const dbKeys = [
@@ -199,7 +341,7 @@ const dbKeys = [
   'full_art',
   'story_spotlight',
   'edhrec_rank',
-  'related_uris' 
+  'related_uris',
 ]
 
 console.log(dbKeys.length)
